Remove and collapse unneeded html
I can't tell whether this is more dickish or less than the people who
write semicolon-less javascript. I suppose the reasons for conforming is
roughly the same, i.e.

* Prevent "hidden" bugs that have to do with missing characters
* Normalize formatting, so the parts that do need endings or semicolons
  don't look wierd.

But overall, I find it less reprehensible, because HTML is supposed to
be a markup language, with the text first, not the XML tree structure.
JS on the other hand, is supposed to look like java, and I guess I just
feel more attached to those semicolons. Or, it's a lot easier to get
burned by JS than HTML.

I dunno about removing quotes around non-spaced attributes though. Maybe
that's over the line. maybe.